config: moved to /app root
