+++ app/src/main/java/com/nbusy/app/data/InMemDB.java
@@ -2,6 +2,7 @@ package com.nbusy.app.data;
 
 import android.os.AsyncTask;
 
+import com.nbusy.app.Config;
 import com.nbusy.app.data.callbacks.CreateProfileCallback;
 import com.nbusy.app.data.callbacks.GetChatMessagesCallback;
 import com.nbusy.app.data.callbacks.GetProfileCallback;
