+++ app/src/main/res/layout/activity_login.xml
@@ -28,6 +28,7 @@
         android:text="Production Mode"
         android:id="@+id/production_mode_button"
         android:layout_gravity="center_horizontal"
-        android:layout_marginTop="25dp" />
+        android:layout_marginTop="25dp"
+        android:visibility="gone"/>
 
 </LinearLayout>
